refactoring

+++ src/main/java/com/codevibe/codevibe/controllers/CarAdvertController.java
@@ -1,7 +1,6 @@
 package com.codevibe.codevibe.controllers;
 
 import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.json.JsonParseException;
 import org.springframework.http.HttpStatus;
@@ -14,13 +13,17 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-
-import com.codevibe.codevibe.models.CarAdvert;
+import com.codevibe.codevibe.dto.CarAdvertDto;
 import com.codevibe.codevibe.services.CarAdvertService;
 
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import jakarta.validation.Valid;
+
+
 @RestController
-//@RequestMapping("/car/adverts")
-public class CarAdvertController {
+public class CarAdvertController
+ {
 
     @Autowired
     private CarAdvertService carAdvertService;
@@ -40,34 +43,32 @@ public String test() {
 }
 
 //mozda napraviti objekt za errore ?
-
-@GetMapping("/car/adverts")
-    public ResponseEntity<?> getAllAdverts(@RequestParam(required = false, defaultValue = "id") String sortBy)
+    @GetMapping("/car/adverts")
+    public ResponseEntity<?> getAllAdverts(@RequestParam(required = false, defaultValue = "id",name ="sortBy") String sortBy)
     {
         try
         {
-            List<CarAdvert> allCarAdverts=carAdvertService.getAllCarAdverts(sortBy);
-            return ResponseEntity.status(HttpStatus.OK).body(allCarAdverts);
+            List<CarAdvertDto> allDtoCarAdverts=carAdvertService.getAllCarAdverts(sortBy);
+            return ResponseEntity.status(HttpStatus.OK).body(allDtoCarAdverts);
         } 
         catch (Exception e) 
         {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Exception caused by internal server error!");
         }
-    
 }
 
   @GetMapping("car/adverts/{id}")
-    public ResponseEntity<?> getAdvertById(@PathVariable Long id) 
+    public ResponseEntity<?> getAdvertById(@PathVariable(name ="id") Long id) 
     {
         try 
         {
-            CarAdvert carAdvert = carAdvertService.getCarAdvertById(id);
-            if (carAdvert!=null) 
+            CarAdvertDto carAdvertDto = carAdvertService.getCarAdvertById(id);
+            if (carAdvertDto!=null) 
             {
-                return ResponseEntity.status(HttpStatus.OK).body(carAdvert);
+                return ResponseEntity.status(HttpStatus.OK).body(carAdvertDto);
             }
-             else 
-             {
+            else 
+            {
                 return ResponseEntity.status(HttpStatus.NOT_FOUND).body("No car advert with the given ID was found.");
             }
         } 
@@ -78,46 +79,47 @@ public String test() {
        
     }
 
+    //ovdje heldnat eror 400 bag request
   @PostMapping("car/advert")
-    public ResponseEntity<?> createCarAdvert(@RequestBody CarAdvert newCarAdvert)
+  @ApiResponses(value = { @ApiResponse(responseCode = "200", description = "Successful"),
+			@ApiResponse(responseCode = "400", description = "This is returned if JSON is invalid or cannot be parsed"),
+			@ApiResponse(responseCode = "422", description = "Validacije?") })
+    public ResponseEntity<?> createCarAdvert(@Valid @RequestBody(required = true) CarAdvertDto newDtoCarAdvert)
      {
         try 
         {
-            // razmisliti gdje stavititi metodu za validaciju, gdje ju smjestiti
-            //if(Prode validaciju)
-            carAdvertService.createCarAdvert(newCarAdvert);
-            return ResponseEntity.status(HttpStatus.CREATED).body(carAdvertService.createCarAdvert(newCarAdvert));
-            /*
-             * else
-             * {
-             * return ResponseEntity.status(HttpStatus.UNPROCESSABLE_ENTITY).body("Validation failed!");
-             * }
-             */
+            CarAdvertDto createdCarAdvertDto=carAdvertService.createCarAdvert(newDtoCarAdvert);
+            if(createdCarAdvertDto!=null)
+            {
+                return ResponseEntity.status(HttpStatus.CREATED).body(createdCarAdvertDto);
+            }
+            else 
+            {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("No car advert with the given ID was found.");
+            }
+          
         }
         //provjeriti kako bad request hendlati
         catch (JsonParseException e)
         {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("This is returned if JSON is invalid or cannot be parsed");
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("ddd");
         } 
         
     }
 
      @PutMapping("car/adverts/{id}")
-    public ResponseEntity<?> updateCarAdvert(@PathVariable Long id, @RequestBody CarAdvert updatedAdvert) 
+    public ResponseEntity<?> updateCarAdvert(@PathVariable(name = "id") Long id, @Valid @RequestBody(required = true) CarAdvertDto updatedAdvertDto) 
     {
-
-        CarAdvert existingAdvertOptional = carAdvertService.getCarAdvertById(id);
-        if (existingAdvertOptional!=null) 
+        CarAdvertDto existingCarAdvert = carAdvertService.getCarAdvertById(id);
+        if (existingCarAdvert!=null) 
         {
-            //if(prođe Validaciju)
-           return ResponseEntity.status(HttpStatus.OK).body(carAdvertService.updateCarAdvert(updatedAdvert));
+           return ResponseEntity.status(HttpStatus.OK).body(carAdvertService.updateCarAdvert(updatedAdvertDto));
            /*
             * else 
             {
                 return ResponseEntity.status(HttpStatus.UNPROCESSABLE_ENTITY).body("Validation failed!");
             }
             */
-           
         }
         else 
         {
@@ -127,14 +129,17 @@ public String test() {
         //jos dodati bad request
     }
 
-    @DeleteMapping("car/adverts/{:id}")
-    public ResponseEntity<?> deleteCarAdvert(@PathVariable Long id)
+    @DeleteMapping("car/adverts/{id}")
+    public ResponseEntity<?> deleteCarAdvert(@PathVariable(name = "id",required = true) Long id)
     {
-        try {
+        CarAdvertDto carAdvertForDelete = carAdvertService.getCarAdvertById(id);
+        if(carAdvertForDelete!=null)
+        {
             carAdvertService.deleteCarAdvert(id);
             return ResponseEntity.status(HttpStatus.NO_CONTENT).body("None");
-        } catch (Exception e) {
-            //kako ovdje provjeriti koji je ecpetion jeli to exception koji je uzrokovan zbog not fount ili nesto drugo?
+        }
+        else 
+        {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("This is returned if a car advert with given id is not found");
         }
     }
